exchange_platform/requirements.txt
amqp
anyio
asgiref
billiard
certifi
click
click-didyoumean
click-plugins
click-repl
Django
django-grappelli
django-redis
djangorestframework
h11
httpcore
httpx
idna
kombu
pillow
psycopg2-binary
prompt_toolkit
python-dateutil
redis
six
sniffio
sqlparse
typing_extensions
tzdata
vine
wcwidth
